refactor: 구간/노선도 호출 구조 변경
